Flask/project/requirements.txt
azure-cognitiveservices-vision-customvision==1.0.0
azure-common==1.1.23
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msrest==0.6.10
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
Pillow==6.1.0
python-dateutil==2.8.0
python-resize-image==1.1.19
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
Werkzeug==0.15.6
